refactor: 레이아웃 UI 이동 완료 (#99)
